requirements.txt
mss>=9.0.1
pynput>=1.7.6
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
anthropic>=0.25.0
openai>=1.25.0
python-dotenv>=1.0.0
Pillow>=10.3.0
python-multipart>=0.0.9
